+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31" >
         <receiver
-            android:name=".view.seatselection.AlarmReceiver"
+            android:name=".view.alarm.AlarmReceiver"
             android:enabled="true"
             android:exported="true" >
         </receiver>
