+++ app/src/main/java/woowacourse/movie/view/alarm/AlarmController.kt
@@ -1,0 +1,42 @@
+package woowacourse.movie.view.alarm
+
+import android.app.AlarmManager
+import android.app.NotificationChannel
+import android.app.NotificationManager
+import android.app.PendingIntent
+import android.content.Context
+
+class AlarmController(
+    private val context: Context
+) {
+
+    init {
+        createChannel()
+    }
+
+    private val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
+
+    fun registerAlarm(pendingIntent: PendingIntent, triggerTime: Long) {
+        alarmManager.set(
+            AlarmManager.RTC_WAKEUP,
+            triggerTime,
+            pendingIntent,
+        )
+    }
+
+    private fun createChannel() {
+        val channel = NotificationChannel(
+            CHANNEL_ID,
+            CHANNEL_NAME,
+            NotificationManager.IMPORTANCE_DEFAULT,
+        )
+        val notificationManager =
+            context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+        notificationManager.createNotificationChannel(channel)
+    }
+
+    companion object {
+        private const val CHANNEL_NAME = "Reservation Notification"
+        const val CHANNEL_ID = "RESERVATION_CHANNEL"
+    }
+}
